Streamline, deduplicate tool implementations
