+++ src/tools/get-avatar-by-id.ts
@@ -3,8 +3,7 @@ import { zodToJsonSchema } from 'zod-to-json-schema';
 import { validateHash } from '../common/utils.js';
 import { DefaultAvatarOption } from '../common/types.js';
 import { Rating } from '../generated/gravatar-api/models/Rating.js';
-import { getUserAgent } from '../config/server-config.js';
-import { GravatarValidationError } from '../common/errors.js';
+import { fetchAvatar } from './avatar-utils.js';
 
 // Schema definition
 export const getAvatarByIdSchema = z.object({
@@ -38,54 +37,6 @@ export const getAvatarByIdTool = {
   description: 'Get the avatar PNG image for a Gravatar profile using an avatar identifier.',
   inputSchema: zodToJsonSchema(getAvatarByIdSchema),
 };
-
-// Direct avatar fetching function
-async function fetchAvatar(params: z.infer<typeof getAvatarByIdSchema>): Promise<Buffer> {
-  if (!validateHash(params.avatarIdentifier)) {
-    throw new GravatarValidationError('Invalid identifier format');
-  }
-
-  // Build avatar URL
-  let url = `https://gravatar.com/avatar/${params.avatarIdentifier}`;
-  const queryParams = new URLSearchParams();
-
-  if (params.size) {
-    queryParams.append('s', params.size.toString());
-  }
-
-  if (params.defaultOption) {
-    queryParams.append('d', params.defaultOption);
-  }
-
-  if (params.forceDefault) {
-    queryParams.append('f', 'y');
-  }
-
-  if (params.rating) {
-    queryParams.append('r', params.rating);
-  }
-
-  // Add query string to URL if there are any parameters
-  const queryString = queryParams.toString();
-  if (queryString) {
-    url += `?${queryString}`;
-  }
-
-  // Fetch the image
-  const response = await fetch(url, {
-    headers: {
-      'User-Agent': getUserAgent(),
-    },
-  });
-
-  if (!response.ok) {
-    throw new GravatarValidationError(`Failed to fetch avatar: ${response.statusText}`);
-  }
-
-  // Convert the response to a buffer
-  const arrayBuffer = await response.arrayBuffer();
-  return Buffer.from(arrayBuffer);
-}
 
 // Tool handler
 export async function handler(params: z.infer<typeof getAvatarByIdSchema>) {
