+++ src/tools/get-interests-by-email.ts
@@ -1,9 +1,8 @@
 import { z } from 'zod';
 import { zodToJsonSchema } from 'zod-to-json-schema';
-import { ExperimentalApi } from '../generated/gravatar-api/apis/ExperimentalApi.js';
-import { createRestApiConfig } from '../config/server-config.js';
 import { validateEmail, generateIdentifierFromEmail } from '../common/utils.js';
 import { GravatarValidationError } from '../common/errors.js';
+import { fetchInterestsById } from './experimental-utils.js';
 
 // Schema definition
 export const getInferredInterestsByEmailSchema = z.object({
@@ -29,16 +28,6 @@ export async function handler(params: z.infer<typeof getInferredInterestsByEmail
   // Generate identifier from email
   const profileIdentifier = generateIdentifierFromEmail(params.email);
 
-  // Use direct API client to get interests by ID
-  const config = createRestApiConfig();
-  const experimentalApi = new ExperimentalApi(config);
-  const interests = await experimentalApi.getProfileInferredInterestsById({
-    profileIdentifier: profileIdentifier,
-  });
-
-  // Extract just the name field from each interest
-  const interestNames = interests.map((interest: { name: string }) => interest.name);
-  return {
-    content: [{ type: 'text', text: JSON.stringify(interestNames, null, 2) }],
-  };
+  // Use shared interests fetching utility
+  return await fetchInterestsById(profileIdentifier);
 }
