+++ src/tools/get-profile-by-email.ts
@@ -1,11 +1,10 @@
 import { z } from 'zod';
 import { zodToJsonSchema } from 'zod-to-json-schema';
-import { ProfilesApi } from '../generated/gravatar-api/apis/ProfilesApi.js';
-import { createRestApiConfig } from '../config/server-config.js';
 import { validateEmail, generateIdentifierFromEmail } from '../common/utils.js';
 import { GravatarValidationError } from '../common/errors.js';
+import { fetchProfileById } from './profile-utils.js';
 
-// Schema definition (moved from service)
+// Schema definition
 export const getProfileByEmailSchema = z.object({
   email: z.string().refine(validateEmail, {
     message: 'Invalid email format.',
@@ -29,14 +28,6 @@ export async function handler(params: z.infer<typeof getProfileByEmailSchema>) {
   // Generate identifier from email
   const profileIdentifier = generateIdentifierFromEmail(params.email);
 
-  // Use direct API client to get profile by ID
-  const config = createRestApiConfig();
-  const profilesApi = new ProfilesApi(config);
-  const profile = await profilesApi.getProfileById({
-    profileIdentifier: profileIdentifier,
-  });
-
-  return {
-    content: [{ type: 'text', text: JSON.stringify(profile, null, 2) }],
-  };
+  // Use shared profile fetching utility
+  return await fetchProfileById(profileIdentifier);
 }
