+++ src/tools/get-profile-by-id.ts
@@ -1,10 +1,9 @@
 import { z } from 'zod';
 import { zodToJsonSchema } from 'zod-to-json-schema';
-import { ProfilesApi } from '../generated/gravatar-api/apis/ProfilesApi.js';
-import { createRestApiConfig } from '../config/server-config.js';
 import { validateHash } from '../common/utils.js';
+import { fetchProfileById } from './profile-utils.js';
 
-// Schema definition (moved from service)
+// Schema definition
 export const getProfileByIdSchema = z.object({
   profileIdentifier: z.string().refine(validateHash, {
     message:
@@ -21,12 +20,5 @@ export const getProfileByIdTool = {
 
 // Tool handler
 export async function handler(params: z.infer<typeof getProfileByIdSchema>) {
-  const config = createRestApiConfig();
-  const profilesApi = new ProfilesApi(config);
-  const profile = await profilesApi.getProfileById({
-    profileIdentifier: params.profileIdentifier,
-  });
-  return {
-    content: [{ type: 'text', text: JSON.stringify(profile, null, 2) }],
-  };
+  return await fetchProfileById(params.profileIdentifier);
 }
